requirements.txt
python-dotenv==1.0.1
semantic-kernel==0.9.3b1
streamlit==1.32.2